refactor: rename `lib/utils` to `utils/randomStrign`
